keyboardContext, InputDebugMsg, compactFormation

+++ shared/input/InputImpulse.cpp
@@ -2,12 +2,12 @@
 
 void Serializer::write(const INPUT_IMPULSE& value, SerializationStream& stream)
 {
-	Serializer::write((int32_t)value, stream);
+	Serializer::write((int8_t)value, stream);
 }
 
 void Serializer::read(INPUT_IMPULSE& value, SerializationStream& stream)
 {
-	int32_t t;
+	int8_t t;
 	Serializer::read(t, stream);
 	value = (INPUT_IMPULSE)t;
 }
